+++ packages/pull-quote/src/pull-quote-content.js
@@ -1,0 +1,14 @@
+import React from "react";
+import PropTypes from "prop-types";
+import { Text } from "react-native";
+import styles from "./styles";
+
+const PullQuoteContent = ({ children }) => (
+  <Text style={styles.content}>{children}</Text>
+);
+
+PullQuoteContent.propTypes = {
+  children: PropTypes.string.isRequired
+};
+
+export default PullQuoteContent;