+++ packages/pull-quote/src/pull-quote-content.web.js
@@ -1,0 +1,3 @@
+import PullQuoteContent from "./styles/responsive";
+
+export default PullQuoteContent;