+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.9"></a>
+## [0.15.9](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.8...@times-components/article-byline@0.15.9) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.15.8"></a>
+## [0.15.8](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.7...@times-components/article-byline@0.15.8) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.15.7"></a>
+## [0.15.7](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.6...@times-components/article-byline@0.15.7) (2018-04-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.15.6"></a>
+## [0.15.6](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.5...@times-components/article-byline@0.15.6) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
 <a name="0.15.5"></a>
 ## [0.15.5](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.4...@times-components/article-byline@0.15.5) (2018-04-24)